Add XPI to `get_release()`
In PR #179 we unified the "update release status" API endpoint to
include XPI, but didn't update the response schema. Additionally,
`shipitscript` queries `/releases/$releasename` to verify that the
release has been properly updated.

* Move all XPI models to `shipit_api.common` (no changes)
* Allow multiple response schemas
* Fix `generate_public_api_schema` paths
* regenerate public API definition
