+++ api/src/shipit_api/admin/__init__.py
@@ -6,7 +6,6 @@
 import os
 
 import backend_common
-import shipit_api.admin.models  # noqa
 import shipit_api.common.models  # noqa
 from shipit_api.admin.worker import cmd
 from shipit_api.common.config import APP_NAME
